feat: ajustes no modelo de ocr

+++ spacerman_ai/machine_models/Friday/commons.py
@@ -20,6 +20,15 @@ class FridayDataGenerator:
         self.files_indexed = []
         self._workfolder = WindowsConfigs.ROOT_DIR / 'img_dst'
         self.df = None
+        self.__sequence = 0
+
+    def reset_sequence(self):
+        self.__sequence = 0
+        return self.__sequence
+
+    def count_sequence(self):
+        self.__sequence += 1
+        return self.__sequence
 
     def set_subdir_as_workdir(self, folder_name):
         """
@@ -63,13 +72,17 @@ class FridayDataGenerator:
             ocr_text = ocr_text.strip().replace('\n', '').replace('"', '').replace("'", '')
             data_dict['text'] = ocr_text
             data_dict['crash_value'] = crash_val
-            self.collected_data.append(data_dict)
-            if verbose:
+            sequence = self.count_sequence()
+            if crash_val < 0:
+                sequence = self.reset_sequence()
+                print(f'Valor lido:>|{ocr_text}|<')
                 print('Dados:')
                 for k, v in data_dict.items():
                     print(k, v, sep=': ')
                 print('txt: ', ocr_clean_text, 'crsh: ', crash_val)
                 print('--')
+            data_dict['sequence'] = sequence
+            self.collected_data.append(data_dict)
 
     def save_collected_data(self, dest_file: str = None) -> None:
         """
@@ -97,17 +110,19 @@ class FridayUtils:
         :param text:
         :return:
         """
-        possible_values = 'I X . K'.split(' ')
+        possible_values = 'X . K'.split(' ')
         text = text.upper().strip().replace(' ', '') \
-            .replace('I', '').replace('L', '').replace('K', '') \
+            .replace('I', '1').replace('L', '1').replace('K', '') \
+            .replace('T', '1').replace('%', '').replace(',', '') \
+            .replace('A', '4').replace('Q', '0').replace('S', '5') \
             .replace('X', '').replace('‘', '').replace('(', '').replace(')', '') \
-            .replace('\n', '').replace('"', '').replace("'", '')
+            .replace('\n', '').replace('"', '').replace("'", '').replace('O', '0')
         if len(text) > 1:
             if text[-1] in possible_values:
                 text = text[0:-1]
         if len(text) == 0:
             print('len 0')
-            text = -1
+            text = -99999999
         return text
 
     def transform_collected_text(self, text, flag=False):
@@ -121,10 +136,10 @@ class FridayUtils:
             result = float(text)
         except ValueError or TypeError:
             print(f'Não convertido: {text}')
-            result = -1
+            result = -999999.99
         except Exception as e:
             print("Erro não tratado: ", e)
-            result = -2
+            result = -9999999.9
         return result
 
     def get_data_from_image_name(self, filename: str) -> dict:
@@ -174,7 +189,7 @@ class FridayUtils:
             raise ValueError('O tem mais valores que o permitido.')
         return splited
 
-    def get_string_from_image(self, img_path: str) -> str:
+    def get_string_from_image(self, img_path: str, flag=False) -> str:
         """
         Usando o tesseract para ler as imagens e coletar os crashses.
         Le o arquivo e apaga
@@ -186,17 +201,24 @@ class FridayUtils:
         # Read image with opencv
         img = cv2.imread(img_path)
         # Convert to gray
+        time_prefix = datetime.now().strftime('%Y%m%d_%H%M%S')
+        thres_file = str(self.ocr_worker / f"{img_filename}_ocr@{time_prefix}__gry_thres.png")
         img = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
+        if flag:
+            thres_file = str(self.ocr_worker / f"{img_filename}_ocr@{time_prefix}__bw_thres.png")
+            (thresh, img) = cv2.threshold(img, 195, 255, cv2.THRESH_BINARY)
         # Apply dilation and erosion to remove some noise
         kernel = np.ones((1, 1), np.uint8)
         img = cv2.dilate(img, kernel, iterations=1)
         img = cv2.erode(img, kernel, iterations=1)
         # Write the image after apply opencv to do some ...
-        thres_file = str(self.ocr_worker / f"{img_filename}_ocr@{datetime.now().strftime('%Y%m%d_%H%M%S')}__thres.png")
         cv2.imwrite(thres_file, img)
         # Recognize text with tesseract for python
         result = pytesseract.image_to_string(Image.open(thres_file))
-        Path(thres_file).unlink()
+        if not len(result) and not flag:
+            result = self.get_string_from_image(img_path=img_path, flag=True)
+        if len(result):
+            Path(thres_file).unlink()
         return result
